Title: 개발팀의 능력, Time: 143ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240213/개발팀의 능력/the-capabilities-of-the-development-team.java
@@ -37,7 +37,7 @@ public class Main {
 		int sum3 = total - (sum1 + sum2);
 
 		if(sum1 == sum2 && sum2 == sum3){
-			return Integer.MAX_VALUE;
+			return (int)Integer.MAX_VALUE;
 		}
 
 		int max = Math.max(Math.max(sum1, sum2), sum3);
